CLEANUP: Fit homomorphisms to presentations
In subgroup lattice, subgroups know an fp version of their image in the
factor, this is used to compute complements via cohomology. When
initializing this list for the subgroups of the radical factor, make sure
that generators chosen for the subgroups align with generators for the
presentations -- this is cleaner, possibly minusculy quicker, and also
eliminates problems that might have crept in prior code that had to use
multiplication with restricted mappings.