kernel: use HaveInterrupt instead of SyIsIntr
Also get rid of user interrupt check in the kernel function PrintObj():
the only ways I found to trigger that in practice first resulted in
the generic ctrl-C handler triggering; only upon using `return;` to
continue from that did the interrupt check in `PrintObj()` trigger,
meaning that the user had to use `return;` twice.

To see an example (using GAP *before* this commit, of course), you
can run this code and interrupt the printing.

    l:=[1..10];;
    for i in [1..5] do l:=List([1..10], i -> ShallowCopy(l)); od;
    PrintObj(l);