re-added the `ViewObj` method for generic fields
This method had been lost in 2012 (see Commit v4.6.0)
when a `ViewString` method had replaced it.
Note that other `ViewObj` methods for the objects in question are
applicable that have a higher rank than the delegation to `ViewString`.
The same happens for many objects, but for the moment I want to fix
the behavior for fields.

(This could have been noticed in `tst/testbugfix/2012-08-14-t00271.tst`.)