kernel: optimize IsVecFFE, make KTNumPlist static
Instead of calling KTNumPlist, or TYPE_OBJ, to check if a plain list is an
FFE vector, do so directly. This can save a lot of overhead if the input is
e.g. a matrix (given as a list-of lists), as we don't try to recursively
determine the type of the sublists.