in configure, run Julia without `startup.jl`
I had put a `print` statement into my `startup.jl` file,
for test purposes, and then GAP's `configure` ran into an error
because the printed message was regarded as the value of `JL_SHARE` etc.

I think it is better not to read `startup.jl` at all in this situation.