kernel: avoid pointer casts increasing alignment
A char pointer might be unaligned, casting it to e.g. an UInt ptr thus
is potentially dangerous and the compiler can be told to warn us about that.
This patch fixes several such warnings in cases where they are easy to avoid,
as we start with a pointer with "big" alignment, than cast it twice, first to
smaller alignment and then again to bigger alignment.

Also add static assertion that verifies that sizeof(StatHeader) is a multiple
of sizeof(Stat), as the above change and other existing code implicitly rely
on that.