kernel: avoid access NLOC_FUNC in PrintFunction if possible
Reorder some of the code to ensure printing a  kernel function does not
needlessly access NLOC_FUNC. This avoids a potential problem with custom
functions that abuse NLOC_FUNC to store alternate data, such as the
"SmallInt Bitfield operations" in intfuncs.c.

Also, don't retain a pointer into a GAP object across calls that might
trigger a GC, such as `Pr`. But note that this code is only used in
HPC-GAP, and there we don't use a moving GC, so in practice this code
was fine anyway.