kernel+buildsys: enable -Wstrict-prototypes
This requires us to mark ObjFunc as an exception via a #pragma. We cannot
currently get rid of this without major effort (and risking to break some
packages).

This in turn made it necessary to add -Wno-pragmas, to prevent GCC from
complaining about the #pragma when compiling in C mode.