Add BijectionsBetweenDirectProductDomainAndRange
BijectionsBetweenDirectProductDomainAndRange returns a record with:
- `positionElement` and `elementPosition` the bijections functions,
- `canSort` whether the components could be sorted.

If `canSort` is true the computed bijective functions are order
preserving.

BijectiveMappingFromDirectProductDomainToRange and
BijectiveMappingFromRangeToDirectProductDomain now use this function to
generate `fun` and `invFun`.

AsList and Enumerator now use this function. If `canSort` was set to true
then the resulting list or enumerator will know that it is sorted.